pyproject.toml
[tool.poetry]
name = "dcodex"
version = "1.0.3"
description = "A software framework for manuscript analysis"
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/rbturnbull/dcodex"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
Django = "^3.2.6"
django-guardian = "^2.4.0"
django-polymorphic = "^3.0.0"
numpy = "^1.21.1"
pandas = "^1.3.1"
scipy = "^1.7.1"
matplotlib = "^3.4.2"
python-Levenshtein = "^0.12.2"
pyxDamerauLevenshtein = "^1.7.0"
beautifulsoup4 = "^4.9.3"
django-admin-sortable2 = "^1.0"
lxml = "^4.6.3"
regex = "^2021.8.3"
django-filer = "^2.0.2"
django-imagedeck = {git = "https://gitlab.unimelb.edu.au/rturnbull/django-imagedeck.git"}
gotoh = {git = "https://github.com/rbturnbull/gotoh.git"}

[tool.poetry.dev-dependencies]
coverage = "^5.5"
black = "^21.9b0"

[build-system]
requires = ["setuptools >= 40.6.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
